+++ packages/orq-rc/src/models/operations/createidentity.ts
@@ -139,7 +139,7 @@ export const CreateIdentityResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-05-04T02:40:56.592Z",
+    "2026-05-04T10:12:56.421Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {